Include major functionalities in DIALS.

+++ edtools/update_dials.py
@@ -11,14 +11,16 @@ from instamatic.formats import write_adsc
 from instamatic.formats import write_tiff
 from instamatic.processing import apply_stretch_correction
 from instamatic.processing.stretch_correction import affine_transform_ellipse_to_circle
-from instamatic.tools import find_beam_center
+from instamatic.tools import find_beam_center, find_subranges
 
 from .utils import parse_args_for_fns
 
 
-def update_dials(fn, wavelength, physical_pixelsize, pixelsize, exposure, phi, osc_angle, name,
-             integrate=False, center=False, stretch=False):
-    shutil.copyfile(fn, fn.with_name("dials_process.bat~"))
+def update_dials(fn, wavelength, physical_pixelsize, pixelsize, exposure, phi, osc_angle, name, axis,
+             integrate=False, center=False, stretch=False, refine=False, symmetry=False, scale=False,
+             merge=False, report=False, export=False):
+    if fn.exists():
+        shutil.copyfile(fn, fn.with_name("dials_process.bat~"))
 
     mrc_folder = fn.parent.parent / 'RED'
     smv_folder = fn.parent
@@ -85,7 +87,43 @@ def update_dials(fn, wavelength, physical_pixelsize, pixelsize, exposure, phi, o
         header['BEAM_CENTER_Y'] = f'{center_y*physical_pixelsize:.4f}'
         header['DENZO_X_BEAM'] = f'{center_x*physical_pixelsize:.4f}'
         header['DENZO_Y_BEAM'] = f'{center_y*physical_pixelsize:.4f}'
-        write_adsc(smv_folder/(img_name+'.img'), img, header)
+        write_adsc(smv_folder/'data'/(img_name+'.img'), img, header)
+
+    scanranges = find_subranges(zip(*img_name_list)[0])
+    scanranges = [int(angle) for angle in scanrages]
+    observed_range = set(scanranges)
+    complete_range = set(range(min(observed_range), max(observed_range) + 1))
+    missing_range = observed_range ^ complete_range
+    scanrange = ' '.join(f'scan_range={i},{j}' for i, j in scanranges)
+    excludeimages = ','.join(str(n) for n in missing_range)
+    rot_x, rot_y, rot_z = rotation_axis_to_xyz(axis, setting='dials')
+    with open(fn, 'w', encoding = 'cp1252') as f:
+        print('@echo off', file=f)
+        print('', file=f)
+        print(f'set scan_range={scanrange}', file=f)
+        print(f'set exclude_images=exclude_images={excludeimages}', file=f)
+        print(f'set rotation_axis=geometry.goniometer.axes={rot_x:.4f},{rot_y:.4f},{rot_z:.4f}', file=f)
+        print('dials.import template=./data/#####.img %rotation_axis%', file=f)
+        print('dials.find_spots imported.expt %scan_range%', file=f)
+        print('dials.index imported.expt strong.refl max_lattices=3', file=f)
+        if refine:
+            print('dials.refine_bravais_settings indexed.expt indexed.refl', file=f)
+            print('dials.refine indexed.expt indexed.refl', file=f)
+        if integrate:
+            print('dials.integrate refined.expt refined.refl nproc=8', file=f)
+        if symmetry:
+            print('dials.symmetry integrated.expt integrated.refl', file=f)
+        if scale:
+            print('dials.scale symmetrized.expt symmetrized.refl', file=f)
+        if merge:
+            print('dials.merge scaled.expt scaled.refl', file=f)
+        if report:
+            print('dials.report scaled.expt scaled.refl', file=f)
+        if export:
+            print('dials.export scaled.expt scaled.refl', file=f)
+    empty = np.zeros_like(img)
+    for n in missing_range:
+        write_adsc(smv_folder/'data'/(f'{n:05d}'+'.img'), empty, header)
 
 def main():
     import argparse
@@ -119,10 +157,28 @@ def main():
                         action="store", type=str, dest="name",
                         help="Add camera name")
 
-    parser.set_defaults(integrate=False,
-                        center=False,
-                        stretch=False,
-                        refine=False,
+    parser.add_argument("-sym", "--symmetry",
+                        action="store", type=bool, dest="symmetry",
+                        help="Check the symmetry of the data")
+
+    parser.add_argument("-meg", "--merge",
+                        action="store", type=bool, dest="merge",
+                        help="Merge the data")
+
+    parser.add_argument("-sca", "--scale",
+                        action="store", type=bool, dest="scale",
+                        help="Scale the data")
+
+    parser.add_argument("-rep", "--report",
+                        action="store", type=bool, dest="report",
+                        help="Report the data")
+
+    parser.add_argument("-exp", "--export",
+                        action="store", type=bool, dest="export",
+                        help="Export the data")
+
+    parser.set_defaults(integrate=False, center=False, stretch=False, refine=False,
+                        symmetry=False, scale=False, merge=False, report=False, export=False
                         name='Medipix3',)
 
 
@@ -141,16 +197,19 @@ def main():
                 wavelength = float(line.split(' ')[1])
             elif 'Exposure' in line:
                 exposure = float(line.split(' ')[1])
+            elif 'Rotation axis' in line:
+                axis = float(line.split(' ')[1])
         lines = open(fn.parent/'RED'/'1.ed3d', "r", encoding = 'cp1252').readlines()
         for line in lines:
             if '00001.mrc' in line:
                 phi = float(line.split()[-1])
 
         print("\033[K", fn, end='\r')  # "\033[K" clears line
-        update_dials(fn, wavelength, physical_pixelsize, pixelsize, exposure, phi, osc_angle, name,
-                     integrate=integrate,
-                     center=center,
-                     stretch=stretch,)
+        (fn.parent/'SMV'/'data').mkdir(parents=True, exist_ok=True)
+        update_dials(fn.parent/'SMV'/'dials_process.bat', wavelength, physical_pixelsize, 
+                     pixelsize, exposure, phi, osc_angle, name, axis,
+                     integrate=integrate, center=center, stretch=stretch, refine=refine,
+                     symmetry=symmetry, scale=scale, merge=merge, report=report, export=export)
 
     print(f"\033[KUpdated {len(fns)} files")
